[fix] 회원가입 페이지 types를 컴포넌트 types로 이동
